Add a couple failing tests of behavior in callbacks and fix them
Previously, dropping a control in a callback, having a callback
add/remove any callback of the same type, or having a callback remove
itself would cause a panic.

One panic was due to destroy callbacks not being registered for a
specific Ihandle. The HashMap holding callbacks would be borrowed to
iterate over it, but since all destroy callbacks were called for every
destroyed control (which is slow besides), the destroy callback would
panic trying to borrow it a second time.

A callback removing itself would try to borrow the list of callbacks to
remove itself from it, causing a panic because the list was already
borrowed to iterate over it.

Now, adding or removing a callback will copy the callback list if it is
already borrowed and make modifications to the copy (copy-on-write).