Fix menus so that dialogs actually keep them from getting destroyed
Turns out that if a dialog already has a child, then it seems
IupAppend(dialog, menu) doesn't set the menu as a child. So then when
the last HandleRc holding a reference to the menu was dropped, it
destroyed the menu, leaving the dialog with a dangling pointer. Now
there's a pretty solid way of tracking non-child references from one
control to another. Even if the dialog never sets the menu as a child,
it should work fine. The only thing that would cause a problem is if
the dialog keeps the menu as a child when it shouldn't.