Add a macro for defining callbacks more easily
I'm trying to avoid macros, but every time I needed to wrap a callback,
I found myself dreading the task. Part of the reason was that it was
complex enough that I couldn't remember how to do it (i.e., it would
take me a long time without copy and paste), part was the
duplication. With the macro, it is more readable, has less duplication,
and I can remember how to use it.