Add set_top_level_margin_and_gap method
This method is useful to set a reasonable default for the margin and
gap. It is intended to be called on a Dialog's direct child. Since the
margin and gap are inheritable, it should cover all children of a
dialog too.

I would prefer to do it automatically, but there's no good place.
Perhaps with specialization, I could specialize append for Dialog, and
check for Hbox, Vbox, and GridBox? Doing it explicitly is nice in a way.