Use try_with() instead of with() to access TLS value
When the process is exiting, depending on the drop order, the TLS values
may have already been dropped. Since this only happens when the thread
is destroyed, the code doesn't really need to run anyway. It would be
nice to keep using with() when this code runs if the thread isn't
getting destroyed, but I can't think of a way to do that.

Fixes #18